split text resource
