+++ .github/workflows/quality.yml
@@ -36,7 +36,7 @@ jobs:
 
           # TODO these values should become zero in the future
           max_large_classes=9
-          max_complex_beans=14
+          max_complex_beans=13
 
           echo "=========================================="
           echo "CODE QUALITY ANALYSIS SUMMARY"
